backend/requirements.txt
fastapi
uvicorn
python-multipart
pytesseract
pillow
opencv-python
pdf2image
numpy
python-multipart